changes naming and restructures grouping of components in folders, refactors feature test - moves interceptor before cy.visit
