+++ cypress/integration/visitorCanNavigateBetweenSPAViews.feature.js
@@ -54,17 +54,19 @@ describe("visitor can navigate between views", () => {
 
   describe("Search View", () => {
     before(() => {
-      cy.visit("/search");
+      cy.visit("/services/search");
     });
     it("is expected to display view subtitle", () => {
       cy.get("[data-cy=header-subtitle]").should("contain.text", "Search");
     });
   });
+
   describe("Services View", () => {
     before(() => {
-      cy.visit("/services");
       cy.intercept("GET", "**/api/services", { fixture: "services_view_section.json" });
+      cy.visit("/services");
     });
+
     it("is expected to display service page ", () => {
       cy.get("[data-cy=service-section]").should("have.length", 6);
       cy.get("[data-cy=service-section]")
